Progresso na implementação do flip-card
