+++ src/app/features/strategic-projects/data-chart/critical-milestones-for-performance/criticalMilestonesForPerformance.component.html
@@ -1,6 +1,5 @@
 <ngx-flip-table
-  [cardTitle]="'Entregas por Status'"
-  [tableContent]="flipTableContent"
+  [cardTitle]="'Marcos Críticos por Desempenho'"
 >
   <ngx-pie-chart-model [data]="chartData" [colors]="chartColors" [height]="150">
   </ngx-pie-chart-model>
