+++ src/app/features/strategic-projects/data-chart/deliveries-by-performance/deliveriesByPerformance.component.html
@@ -1,5 +1,5 @@
 <ngx-flip-table
-  [cardTitle]="'Entregas por Status'"
+  [cardTitle]="'Entregas por Desempenho'"
   [tableContent]="flipTableContent"
 >
   <ngx-pie-chart-model [data]="chartData" [colors]="chartColors" [height]="150">
