+++ src/app/features/strategic-projects/data-chart/projects-by-status/projectsByStatus.component.html
@@ -1,6 +1,5 @@
 <ngx-flip-table
-  [cardTitle]="'Entregas por Status'"
-  [tableContent]="flipTableContent"
+  [cardTitle]="'Projetos por Status'"
 >
   <ngx-pie-chart-model [data]="chartData" [colors]="chartColors" [height]="150">
   </ngx-pie-chart-model>
