+++ src/app/features/strategic-projects/data-chart/risks-by-classification/risksByClassification.component.html
@@ -1,6 +1,5 @@
 <ngx-flip-table
-  [cardTitle]="'Entregas por Status'"
-  [tableContent]="flipTableContent"
+  [cardTitle]="'Riscos por Importância'"
 >
   <ngx-pie-chart-model [data]="chartData" [colors]="chartColors" [height]="150">
   </ngx-pie-chart-model>
