+++ src/app/features/strategic-projects/strategicProjects.component.html
@@ -272,7 +272,7 @@
       </div>
 
       <div class="col-xxxl-6 col-xxl-4 col-md-6 pr-1 pl-1 mb-2">
-        <div class="card">
+        <!-- <div class="card">
           <div class="d-flex col-12 mt-1 mb-1">
             <span class="text-card">Entregas por Desempenho</span>
           </div>
@@ -282,11 +282,12 @@
               <ngx-deliveries-by-performace [filter]="finalFilter"></ngx-deliveries-by-performace>
             </div>
           </div>
-        </div>
+        </div> -->
+        <ngx-deliveries-by-performace [filter]="finalFilter"></ngx-deliveries-by-performace>
       </div>
 
       <div class="col-xxxl-6 col-xxl-4 col-md-6 pr-1 pl-1 mb-2">
-        <div class="card">
+        <!-- <div class="card">
           <div class="d-flex col-12 mt-1 mb-1">
             <span class="text-card">Projetos por Status</span>
           </div>
@@ -296,11 +297,12 @@
               <ngx-projects-by-status [filter]="finalFilter"></ngx-projects-by-status>
             </div>
           </div>
-        </div>
+        </div> -->
+        <ngx-projects-by-status [filter]="finalFilter"></ngx-projects-by-status>
       </div>
 
       <div class="col-xxxl-6 col-xxl-4 col-md-6 pr-1 pl-1 mb-2">
-        <div class="card">
+        <!-- <div class="card">
           <div class="d-flex col-12 mt-1 mb-1">
             <span class="text-card">Marcos Críticos por Desempenho</span>
           </div>
@@ -310,11 +312,12 @@
               <ngx-critical-milestones-for-performance [filter]="finalFilter"></ngx-critical-milestones-for-performance>
             </div>
           </div>
-        </div>
+        </div> -->
+        <ngx-critical-milestones-for-performance [filter]="finalFilter"></ngx-critical-milestones-for-performance>
       </div>
 
       <div class="col-xxxl-6 col-xxl-4 col-md-6 pr-1 pl-1 mb-2">
-        <div class="card">
+        <!-- <div class="card">
           <div class="d-flex col-12 mt-1 mb-1">
             <span class="text-card">Entregas por Tipo</span>
           </div>
@@ -323,11 +326,12 @@
               <ngx-deliveries-by-type [filter]="finalFilter"></ngx-deliveries-by-type>
             </div>
           </div>
-        </div>
+        </div> -->
+        <ngx-deliveries-by-type [filter]="finalFilter"></ngx-deliveries-by-type>
       </div>
 
       <div class="col-xxxl-6 col-xxl-4 col-md-6 pr-1 pl-1 mb-2">
-        <div class="card">
+        <!-- <div class="card">
           <div class="d-flex col-12 mt-1 mb-1">
             <span class="text-card">Riscos por Importância</span>
           </div>
@@ -337,7 +341,8 @@
               <ngx-risks-by-classification [filter]="finalFilter"></ngx-risks-by-classification>
             </div>
           </div>
-        </div>
+        </div> -->
+        <ngx-risks-by-classification [filter]="finalFilter"></ngx-risks-by-classification>
       </div>
     </div>
