+++ src/app/features/strategic-projects/strategicProjects.module.ts
@@ -3,11 +3,11 @@ import { NgModule } from '@angular/core';
 import { StrategicProjectsComponent } from './strategicProjects.component';
 import { NbCardModule, NbLayoutModule, NbSelectModule } from '@nebular/theme';
 import { DeliveriesByStatusComponent } from './data-chart/deliveries-by-status/deliveriesByStatus.component';
-import { CriticalMilestonesForPerformanceComponent } from './data-chart/criticalMilestonesForPerformace.component';
-import { DeliveriesByPerformaceComponent } from './data-chart/deliveriesByPerformace.component';
-import { DeliveriesByTypeComponent } from './data-chart/deliveriesByType.component';
-import { ProjectsByStatusComponent } from './data-chart/projectsByStatus.component';
-import { RisksByClassificationComponent } from './data-chart/risksByClassification.component';
+import { CriticalMilestonesForPerformanceComponent } from './data-chart/critical-milestones-for-performance/criticalMilestonesForPerformace.component';
+import { DeliveriesByPerformaceComponent } from './data-chart/deliveries-by-performance/deliveriesByPerformace.component';
+import { DeliveriesByTypeComponent } from './data-chart/deliveries-by-type/deliveriesByType.component';
+import { ProjectsByStatusComponent } from './data-chart/projects-by-status/projectsByStatus.component';
+import { RisksByClassificationComponent } from './data-chart/risks-by-classification/risksByClassification.component';
 import { NgxEchartsModule } from 'ngx-echarts';
 import { AccumulatedInvestmentComponent } from './data-chart/accumulatedInvestment.component';
 import { DeliveriesBySelectedComponent } from './data-chart/deliveriesBySelected.component';
